#150: split updating logic from main Browscap class
