+++ src/Browscap.php
@@ -29,15 +29,8 @@
 
 namespace BrowscapPHP;
 
-use Browscap\Generator\BuildGenerator;
-use Browscap\Helper\CollectionCreator;
-use Browscap\Writer\Factory\PhpWriterFactory;
 use BrowscapPHP\Cache\BrowscapCache;
 use BrowscapPHP\Cache\BrowscapCacheInterface;
-use BrowscapPHP\Exception\FetcherException;
-use BrowscapPHP\Helper\Converter;
-use BrowscapPHP\Helper\Filesystem;
-use BrowscapPHP\Helper\IniLoader;
 use BrowscapPHP\Helper\Quoter;
 use BrowscapPHP\Parser\ParserInterface;
 use Psr\Log\LoggerInterface;
@@ -87,19 +80,6 @@ class Browscap
      * @var @var \Psr\Log\LoggerInterface
      */
     private $logger = null;
-
-    /**
-     * Options for the updater. The array should be overwritten,
-     * containing all options as keys, set to the default value.
-     *
-     * @var array
-     */
-    private $options = array();
-
-    /**
-     * @var \BrowscapPHP\Helper\IniLoader
-     */
-    private $loader = null;
 
     /**
      * Set theformatter instance to use for the getBrowser() result
@@ -236,40 +216,6 @@ class Browscap
     }
 
     /**
-     * Sets multiple loader options at once
-     *
-     * @param array $options
-     *
-     * @return \BrowscapPHP\Browscap
-     */
-    public function setOptions(array $options)
-    {
-        $this->options = $options;
-
-        return $this;
-    }
-
-    /**
-     * @return \BrowscapPHP\Helper\IniLoader
-     */
-    public function getLoader()
-    {
-        if (null === $this->loader) {
-            $this->loader = new IniLoader();
-        }
-
-        return $this->loader;
-    }
-
-    /**
-     * @param \BrowscapPHP\Helper\IniLoader $loader
-     */
-    public function setLoader(IniLoader $loader)
-    {
-        $this->loader = $loader;
-    }
-
-    /**
      * parses the given user agent to get the information about the browser
      *
      * if no user agent is given, it uses {@see \BrowscapPHP\Helper\Support} to get it
@@ -288,6 +234,11 @@ class Browscap
             $userAgent = $support->getUserAgent();
         }
 
+        if (null === $this->getCache()->getVersion()) {
+            // there is no active/warm cache available
+            throw new Exception('there is no active cache available, please run the update commend');
+        }
+
         // try to get browser data
         $formatter = $this->getParser()->getBrowser($userAgent);
 
@@ -298,269 +249,5 @@ class Browscap
         }
 
         return $formatter->getData();
-    }
-
-    /**
-     * reads and parses an ini file and writes the results into the cache
-     *
-     * @param  string $iniFile
-     *
-     * @throws \BrowscapPHP\Exception
-     */
-    public function convertFile($iniFile)
-    {
-        $loader = new IniLoader();
-
-        try {
-            $loader->setLocalFile($iniFile);
-        } catch (Helper\Exception $e) {
-            throw new Exception('an error occured while setting the local file', 0, $e);
-        }
-
-        try {
-            $iniString = $loader->load();
-        } catch (Helper\Exception $e) {
-            throw new Exception('an error occured while converting the local file into the cache', 0, $e);
-        }
-
-        $this->convertString($iniString);
-    }
-
-    /**
-     * reads and parses an ini string and writes the results into the cache
-     *
-     * @param string $iniString
-     */
-    public function convertString($iniString)
-    {
-        $cachedVersion = $this->getCache()->getItem('browscap.version', false, $success);
-        $converter     = new Converter($this->getLogger(), $this->getCache());
-
-        $this->storeContent($converter, $iniString, $cachedVersion);
-    }
-
-    /**
-     * fetches a remote file and stores it into a local folder
-     *
-     * @param string $file       The name of the file where to store the remote content
-     * @param string $remoteFile The code for the remote file to load
-     *
-     * @throws \BrowscapPHP\Exception\FetcherException
-     * @throws \BrowscapPHP\Helper\Exception
-     */
-    public function fetch($file, $remoteFile = IniLoader::PHP_INI)
-    {
-        if (null === ($cachedVersion = $this->checkUpdate($remoteFile))) {
-            // no newer version available
-            return;
-        }
-
-        $this->getLoader()
-            ->setRemoteFilename($remoteFile)
-            ->setOptions($this->options)
-            ->setLogger($this->getLogger())
-        ;
-
-        $this->getLogger()->debug('started fetching remote file');
-
-        try {
-            $content = $this->getLoader()->load();
-        } catch (Helper\Exception $e) {
-            throw new FetcherException('an error occured while fetching remote data', 0, $e);
-        }
-
-        if (false === $content) {
-            $error = error_get_last();
-            throw FetcherException::httpError($this->getLoader()->getRemoteIniUrl(), $error['message']);
-        }
-
-        $this->getLogger()->debug('finished fetching remote file');
-        $this->getLogger()->debug('started storing remote file into local file');
-
-        $content = $this->sanitizeContent($content);
-
-        $converter  = new Converter($this->getLogger(), $this->getCache());
-        $iniVersion = $converter->getIniVersion($content);
-
-        if ($iniVersion > $cachedVersion) {
-            $fs = new Filesystem();
-            $fs->dumpFile($file, $content);
-        }
-
-        $this->getLogger()->debug('finished storing remote file into local file');
-    }
-
-    /**
-     * fetches a remote file, parses it and writes the result into the cache
-     *
-     * if the local stored information are in the same version as the remote data no actions are
-     * taken
-     *
-     * @param string      $remoteFile The code for the remote file to load
-     * @param string|null $buildFolder
-     * @param int|null    $buildNumber
-     *
-     * @throws \BrowscapPHP\Exception\FileNotFoundException
-     * @throws \BrowscapPHP\Helper\Exception
-     * @throws \BrowscapPHP\Exception\FetcherException
-     */
-    public function update($remoteFile = IniLoader::PHP_INI, $buildFolder = null, $buildNumber = null)
-    {
-        $this->getLogger()->debug('started fetching remote file');
-
-        $converter = new Converter($this->getLogger(), $this->getCache());
-
-        if (class_exists('\Browscap\Browscap')) {
-            $resourceFolder = 'vendor/browscap/browscap/resources/';
-
-            if (null === $buildNumber) {
-                $buildNumber = (int)file_get_contents('vendor/browscap/browscap/BUILD_NUMBER');
-            }
-
-            if (null === $buildFolder) {
-                $buildFolder = 'resources';
-            }
-
-            $buildFolder .= '/browscap-ua-test-'.$buildNumber;
-            $iniFile     = $buildFolder.'/full_php_browscap.ini';
-
-            mkdir($buildFolder, 0777, true);
-
-            $writerCollectionFactory = new PhpWriterFactory();
-            $writerCollection        = $writerCollectionFactory->createCollection($this->getLogger(), $buildFolder);
-
-            $buildGenerator = new BuildGenerator($resourceFolder, $buildFolder);
-            $buildGenerator
-                ->setLogger($this->getLogger())
-                ->setCollectionCreator(new CollectionCreator())
-                ->setWriterCollection($writerCollection)
-                ->run($buildNumber, false)
-            ;
-
-            $converter
-                ->setVersion($buildNumber)
-                ->storeVersion()
-                ->convertFile($iniFile)
-            ;
-
-            $filesystem = new Filesystem();
-            $filesystem->remove($buildFolder);
-        } else {
-            if (null === ($cachedVersion = $this->checkUpdate($remoteFile))) {
-                // no newer version available
-                return;
-            }
-
-            $this->getLoader()
-                ->setRemoteFilename($remoteFile)
-                ->setOptions($this->options)
-                ->setLogger($this->getLogger())
-            ;
-
-            try {
-                $content = $this->getLoader()->load();
-            } catch (Helper\Exception $e) {
-                throw new FetcherException('an error occured while loading remote data', 0, $e);
-            }
-
-            if (false === $content) {
-                $internalLoader = $this->getLoader()->getLoader();
-                $error          = error_get_last();
-
-                throw FetcherException::httpError($internalLoader->getUri(), $error['message']);
-            }
-
-            $this->getLogger()->debug('finished fetching remote file');
-
-            $this->storeContent($converter, $content, $cachedVersion);
-        }
-    }
-
-    /**
-     * checks if an update on a remote location for the local file or the cache
-     *
-     * @param string $remoteFile
-     *
-     * @return int|null The actual cached version if a newer version is available, null otherwise
-     * @throws \BrowscapPHP\Helper\Exception
-     * @throws \BrowscapPHP\Exception\FetcherException
-     */
-    public function checkUpdate($remoteFile = IniLoader::PHP_INI)
-    {
-        $success       = null;
-        $cachedVersion = $this->getCache()->getItem('browscap.version', false, $success);
-
-        if (!$cachedVersion) {
-            // could not load version from cache
-            $this->getLogger()->info('there is no cached version available, please update from remote');
-
-            return 0;
-        }
-
-        $this->getLoader()
-            ->setRemoteFilename($remoteFile)
-            ->setOptions($this->options)
-            ->setLogger($this->getLogger())
-        ;
-
-        try {
-            $remoteVersion = $this->getLoader()->getRemoteVersion();
-        } catch (Helper\Exception $e) {
-            throw new FetcherException('an error occured while checking remote version', 0, $e);
-        }
-
-        if (!$remoteVersion) {
-            // could not load remote version
-            $this->getLogger()->info('could not load version from remote location');
-
-            return 0;
-        }
-
-        if ($cachedVersion && $remoteVersion && $remoteVersion <= $cachedVersion) {
-            // no newer version available
-            $this->getLogger()->info('there is no newer version available');
-
-            return null;
-        }
-
-        $this->getLogger()->info(
-            'a newer version is available, local version: ' . $cachedVersion . ', remote version: ' . $remoteVersion
-        );
-
-        return (int) $cachedVersion;
-    }
-
-    /**
-     * @param string $content
-     *
-     * @return mixed
-     */
-    private function sanitizeContent($content)
-    {
-        // replace everything between opening and closing php and asp tags
-        $content = preg_replace('/<[?%].*[?%]>/', '', $content);
-
-        // replace opening and closing php and asp tags
-        return str_replace(array('<?', '<%', '?>', '%>'), '', $content);
-    }
-
-    /**
-     * reads and parses an ini string and writes the results into the cache
-     *
-     * @param \BrowscapPHP\Helper\Converter $converter
-     * @param string                        $content
-     * @param int|null                      $cachedVersion
-     */
-    private function storeContent(Converter $converter, $content, $cachedVersion)
-    {
-        $iniString  = $this->sanitizeContent($content);
-        $iniVersion = $converter->getIniVersion($iniString);
-
-        if (!$cachedVersion || $iniVersion > $cachedVersion) {
-            $converter
-                ->storeVersion()
-                ->convertString($iniString)
-            ;
-        }
     }
 }
